{{cookiecutter.project_slug}}/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Rodrigo Cortez"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
django-unfold="0.31.0"
psycopg = {extras = ["binary"], version = "^3.1.16"}
djangorestframework = "^3.14.0"
drf-spectacular = "^0.27.0"
gdal = "3.6.2"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
django-oauth-toolkit = "^1.5.0"

[tool.ruff]
fix = true
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
exclude = [
    "**/migrations"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"